refactor(tree): make the age integer in place of double

+++ src/main/java/com/wora/citronix/tree/domain/Level.java
@@ -3,7 +3,7 @@ package com.wora.citronix.tree.domain;
 public enum Level {
     OLD, MATURE, YOUNG;
 
-    public static Level fromAge(Double age) {
+    public static Level fromAge(int age) {
         if (age <= 3) return YOUNG;
         else if (age > 3 && age <= 10) return MATURE;
         else return OLD;
